Add mapbox and pollinations.ai API's to the app
I am using the user input on the search box as the query to mapbox API.
Mapbox API is returning 5 location suggestions based on that query.
I am using 3 properties from the return of the visual crossing API:
resolvedAddress, icon and conditions as query to the pollinations.ai API.
pollinations.ai API returns a generated image that is being load on the
page.